Move memory resource to command buffer pool
This is an attempt to reduce the number of memory resources
and have only one per command buffer pool.  This basically does not
work well because the memory resource can't be reset unless all the
command buffers in the pool get reset at the same time, which does not
usually happen.

Change-Id: I1c44dc1040d345ae233cc5f5d8409e3504aa0f99